confirm vm interface
