+++ src/main/java/com/alibaba/qlexpress4/runtime/instruction/ReturnInstruction.java
@@ -2,7 +2,6 @@ package com.alibaba.qlexpress4.runtime.instruction;
 
 import com.alibaba.qlexpress4.QLOptions;
 import com.alibaba.qlexpress4.exception.ErrorReporter;
-import com.alibaba.qlexpress4.runtime.Parameters;
 import com.alibaba.qlexpress4.runtime.QRuntime;
 
 /**
@@ -18,7 +17,7 @@ public class ReturnInstruction extends QLInstruction {
     }
 
     @Override
-    public void execute(Parameters parameters, QRuntime qRuntime, QLOptions qlOptions) {
+    public void execute(QRuntime qRuntime, QLOptions qlOptions) {
 
     }
 }
